+++ lib/alf-engine/alf/engine/aggregate.rb
@@ -29,12 +29,13 @@ module Alf
       def initialize(operand, summarization)
         @operand = operand
         @summarization = summarization
+        @handle = Tools::TupleHandle.new
       end
 
       # (see Cog#each)
       def each
         agg = operand.inject(@summarization.least) do |memo,tuple|
-          @summarization.happens(memo, tuple)
+          @summarization.happens(memo, @handle.set(tuple))
         end
         yield @summarization.finalize(agg)
       end